fixed minor signal and ioctl bugs

+++ toolchain/newlib-sys-seaos/ctl.c
@@ -20,7 +20,7 @@ int fcntl(int des, int cmd, ...)
 	va_list args;
 	va_start(args, cmd);
 	
-	int ret = syscall(SYS_FCNTL, des, cmd, va_arg(args, int), 0, 0);
+	int ret = syscall(SYS_FCNTL, des, cmd, va_arg(args, long), 0, 0);
 	
 	va_end(args);
 	if(ret < 0) {
@@ -30,7 +30,7 @@ int fcntl(int des, int cmd, ...)
 	return ret;
 }
 
-int ioctl(int fp, int cmd, int arg)
+int ioctl(int fp, int cmd, long arg)
 {
 	int ret = syscall(SYS_IOCTL, fp, cmd, arg, 0, 0);
 	if(ret < 0) {
